Hopefully sped up compile time (52.49 seconds on my machine) Got rid of all warnings in the console

+++ nova-engine/Src/Core/Signal.h
@@ -9,7 +9,7 @@ public:
 
 	int Connect(Callback cb) {
 		cpp_listeners.push_back(cb);
-		return cpp_listeners.size() - 1;
+		return static_cast<int>(cpp_listeners.size()) - 1;
 	}
 
 	void Emit(Args... args) {
